requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9.11
celery>=5.3.6
redis>=5.0.3
django-filter>=23.5
requests>=2.32.5
groq>=0.9.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
django-cors-headers>=4.3
